scripts/youtube-automation/requirements.txt
google-auth-oauthlib==1.3.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-api-python-client==2.104.0
youtube-transcript-api==0.6.3
